+++ src/core/maps/js/mapMarker.js
@@ -108,7 +108,6 @@ function removeMapMarker (layerId) {
     });
 
     markerLayer.getSource().clear();
-
 }
 
 /**